+++ src/main/java/com/bendude56/goldenapple/lock/LockListener.java
@@ -1,4 +1,4 @@
-package com.bendude56.goldenapple.listener;
+package com.bendude56.goldenapple.lock;
 
 import java.sql.SQLException;
 import java.util.logging.Level;
