+++ src/main/java/com/bendude56/goldenapple/lock/LockModuleLoader.java
@@ -4,7 +4,6 @@ import com.bendude56.goldenapple.CommandManager;
 import com.bendude56.goldenapple.ModuleLoader;
 import com.bendude56.goldenapple.commands.AutoLockCommand;
 import com.bendude56.goldenapple.commands.LockCommand;
-import com.bendude56.goldenapple.listener.LockListener;
 import com.bendude56.goldenapple.permissions.PermissionManager;
 
 public class LockModuleLoader extends ModuleLoader {
